Indian_Stock_Market_Analysis/requirements.txt
google-generativeai
yfinance
pandas
pandas_ta
python-dotenv
rich
textblob
